Recompiled to check for Lazarus 1.0 RC2 [now showing 1.0.1.2 in about box].
Tidied up the logic that checks tah a reminder has been set in the futire
date.

If a system shut down is chosen in reminder or countfdown and the
application is being displayed in the system tray, swith to mornam mode -
allows the abore button to be selected.

Added a display to the status bar that indicates the status of the cap,
numlock & scroll lock [in real time].