Although all options are written to the xml file has strings, amended the Options class so that all options are presented to lazKlock  has native type i..e. integers and Booleans.  The conversions are not his within the options class nad not the main program.
Changed the logic in formOptions, wgen the options form loads a backup
options record is created and this is amended, it is only copied to the
main options records if OK is clicked.  If cancel is click the
amendements are discarded.

Added the logic to uKlcoUtils to amend the registry for startinf the app
with windows.

Removed self from vairable referances in uFuzztTime - not usre why they
where there.