formKlock - Moved the loading of Sticky Notes and memo to formload from formShow, this should stop duplicate loading of Stick notes when switching from another klock back to the main Klock.
formOptions, uOptions - added options for Sticky Note Colour and Font.
                        added options for Memo default password, use default
                        password and Time out.

formInfo, uInfoUtils - Added Moon stuff to Info menu.  This includes a graphic
of the moon phase and procedure to calculate the moon phase and illuminance.

uFuzzyTime - Change the algorithm for calculating the Julian Date - this looks
more accurate.

formStickNote - Amended to use default colour and font on create.