Added more Moon info to the Moon info page - all from the DelphiMoon component.
Added a Sun info page - with all info coming from the DephiMoon component.

Added the user options for Longitude and Latitude - used in Sun/Moon rise and setting times.
The defaults are for 10 Downing Street, London :-)

Changed Julian time in uFuzzyTie to use the function from TZPaascal, instead of a calculation.

Changed Easter date in uInfoUtils to use the function from TZPaascal, instead of a calculation.
Aslo changed the moon stuff to use various function from Delphimoon instead of calculations.

uKlockUtils also uses the function for Easter Sunday instead of a calling the one in uInfoUtils.