Tidied up the font handling.  The font work flow needed three entries - one for the file name, one for the name to load the font and a third friendly version for the drop down menus.  All three where hard coded.
This has been changed so that the font directory is scanned and all .ttf
files are loaded.  This directory now contains two text files that contain
the name to load the font [fontNames.txt] and a list of friendly names
[fontTypes.txt].  These text files need to be amended when a new font is
added.  But, this means a new font can be added without re-compiling Klock.

Added the option to include the font directory to the archive.