I found the forms propertie FormStyle whch can be set to fsSystemStayOnTop. Amended the five other Klock [Analogue Klock, LED Klock, Floating Text Klock, Binary Klock and Small Text Klock] to be able to stay on top, this can be either selected from each Klocks pop up menu or set in the user options.
I think Floating text Klock in Gushing Meadow font in red with stay on top enabled looks cool.

Added a plash screen to Klock [formSplashScreen].
This is called when the program starts and closes.  The splash screen just echo entries being made to the log file.  This is useful if a few fonts are being loaded as this slows down both starting and closing Klock.