The current Reminder tab had a couple [many] bugs and did not really work and needed re-writing really.  Also the naming of the Reminder and Event tabs was not really correct.
So, I have deleted the Reminder tab and all associated code.
I have renamed the Events tab to Reminder - This tab is about a single one off
prompt at a given time so suits the name Reminder and not Events.

A new Events tab will be written, hopefully more suited to the job.