Switched on Hi-Dpi aware.
Added two new units - uEvent.pas and uEvents.pas.
Events is a little like a memo, but with a date.  These can be one off or
recurring and Klock will alert you when they due.
All of the GUI element for Events in place.
Code for the New, Clear and add buttons written.
Need to test, before continuing