Created EventTypes.txt, which is loaded at start up to provides the different event types - will allow customisation later.
Corrected a bug with setting the event buttons in frmKlock.
Corrected a bug where the event type was not being passed out of frmKlock.
Tidied up the event code in frmKlock.

Corrected a bug where the event type was not being correctly saved in
uEvent.

Amended the amend procedure in uEvents if handle event types correctly