formOptions - added option to save achieve with either relative or absolute pathnames.  The option is save in user options. uArchiveOUtils - amended the saveArchive and LoadArchive routines to cater for either relative or absolute pathnames.
formEvent - added a second button to event - can now either be mute or
acknowledge.  It mute is clicked the event is cleared and will be displayed next
cycle, it acknowledge is clicked then the event is cleared and the
acknowledge flag on the event is set.  When the acknowledge flag is set, the
event will not be displayed until the next aged band is reached.

uEvents - Corrected bug with displaying age in years.