Added time to uevent.pas and uevents.pas. Amended both, so that the new time fields if included in save, new and amend routines.
Also amended the save and load routines in uevent.pas, so that the boolean
fields are saved and read as strings.  They are then converted to booleans
to be used within Klock.  It seems that booleans can't be read and written
natively.

Corrected a bug in uKlockUtils.KillOtherKlocks, on how it checked on if any other Klocks was running.  Used to use if visible, now uses if assigned.