There was a problem calling the about form from the extra klocks.  Because the about form is not auto-created any more, .show does not work.  Amended each klock to first create the form, show the form and then free the form.
In formScrollingKlock.pas moved the setDisplay from the timer routine to the menu click routine.  Also and the event time to the different routines and make sure the seconds where zero.

Added time to the CSV create in uEvents.pas

Corrected a problem in the read/write of booleans in uEvent.pas.

Corrected a problems with reading the screen position in user options for both
floatings Text Klock and scrolling Text Klcok.

Removal of several debug statements.